[HBASE-7366] Make old logs directory have a subdirectory structure
Author: amirshim

Summary:
To speed up replication, it is useful to have each regionserver's old logs in a separate directory, so we don't have to do LS on a 300,000 file directory (namenode doesn't like that).  So we're changing the oldlogs directory structure to contain subdirectories of the format: "10.159.11.45%3A60020", for example:
/TITAN029-ASH3-HBASE/.oldlogs/10.159.11.45%3A60020/10.159.11.45%3A60020.1355452542203

Test Plan:
Unit tests - passed.

Run on cluster, and make sure that files are being moved to the correct subdirectories, and being deleted after the retention time has elapsed.

Reviewers: liyintang, nspiegelberg, mycnyc

Reviewed By: liyintang

CC: hbase-eng@, davejwatson

Differential Revision: https://phabricator.fb.com/D659154

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1423101 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/master/OldLogsCleaner.java
@@ -91,32 +91,48 @@ public class OldLogsCleaner extends Chore {
     }
   }
 
+  /**
+   * Delete log files directories recursively.
+   * @param files The list of files/directories to traverse.
+   * @param deleteCountLeft Max number of files to delete
+   * @param maxDepth Max Directory depth to recurse
+   * @return Number of files left to delete (deleteCountLeft - number deleted)
+   * @throws IOException
+   */
+  private int cleanFiles(FileStatus[] files, int deleteCountLeft,
+                         int maxDepth) throws IOException {
+    if (files == null || files.length == 0) return deleteCountLeft;
+    if (maxDepth <= 0) {
+      LOG.warn("Old Logs directory structure is too deep: " + files[0].getPath());
+      return deleteCountLeft;
+    }
+    for (FileStatus file : files) {
+      if (deleteCountLeft <= 0) return 0; // we don't have anymore to delete
+      if (file.isDir()) {
+        deleteCountLeft = cleanFiles(this.fs.listStatus(file.getPath()),
+                                     deleteCountLeft, maxDepth - 1);
+        continue;
+      }
+      Path filePath = file.getPath();
+      if (HLog.validateHLogFilename(filePath.getName())) {
+        if (logCleaner.isLogDeletable(filePath) ) {
+          this.fs.delete(filePath, true);
+          deleteCountLeft--;
+        }
+      } else {
+        LOG.warn("Found a wrongly formatted file: "
+            + file.getPath().getName());
+        this.fs.delete(filePath, true);
+        deleteCountLeft--;
+      }
+    }
+    return deleteCountLeft;
+  }
+
   @Override
   protected void chore() {
     try {
-      FileStatus[] files = this.fs.listStatus(this.oldLogDir);
-      if (files == null) {
-        // We don't have any files to process.
-        return;
-      }
-      int nbDeletedLog = 0;
-      for (FileStatus file : files) {
-        Path filePath = file.getPath();
-        if (HLog.validateHLogFilename(filePath.getName())) {
-          if (logCleaner.isLogDeletable(filePath) ) {
-            this.fs.delete(filePath, true);
-            nbDeletedLog++;
-          }
-        } else {
-          LOG.warn("Found a wrongly formated file: "
-              + file.getPath().getName());
-          this.fs.delete(filePath, true);
-          nbDeletedLog++;
-        }
-        if (nbDeletedLog >= maxDeletedLogs) {
-          break;
-        }
-      }
+      cleanFiles(this.fs.listStatus(this.oldLogDir), maxDeletedLogs, 2);
     } catch (IOException e) {
       e = RemoteExceptionHandler.checkIOException(e);
       LOG.warn("Error while cleaning the logs", e);
